requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<0.28
django-filter>=24.2
